!refacto(resources): introduce UiStyleFile class
have a python object to represent the style and move some of the function in browser implying style to that new class.

the browser module is now simpler and only handle paths.

the current implementation have a non-functional font loading system, that need to be fixed later.

improve code style in app.py